requirements.txt
Flask==3.0.3
playwright==1.44.0
python-dotenv==1.0.1
gunicorn==22.0.0
Pillow==10.3.0
pyzbar==0.1.9
requests==2.31.0
